fix IllegalArgumentException reported by François Task #2 - SourceCitationEditorPanel - java.lang.IllegalArgumentException

+++ AncestrisCore/modules.editors.genealogyeditor/src/ancestris/modules/editors/genealogyeditor/panels/SourceCitationEditorPanel.java
@@ -483,7 +483,7 @@ public class SourceCitationEditorPanel extends javax.swing.JPanel {
 
         if (sourcesTableDialog.show() == DialogDescriptor.OK_OPTION) {
             if (sourcesTablePanel.getSelectedSource() != null) {
-                if (mReferencedSource != null) {
+                if (mReferencedSource != null && mReferencedSource.isConnected()) {
                     ((PropertySource) mSourceCitation).unlink();
                 }
                 mReferencedSource = sourcesTablePanel.getSelectedSource();
